Fixed decimal bug for odds

+++ Mefistofeles.Repositories/TeamRepository.cs
@@ -17,7 +17,7 @@ namespace Mefistofeles.Repositories
             DynamicParameters parameter = new DynamicParameters();
 
             parameter.Add("@Name", team.Name, DbType.String, ParameterDirection.Input);
-            parameter.Add("@Odds", team.Odds, DbType.Int64, ParameterDirection.Input);
+            parameter.Add("@Odds", team.Odds, DbType.Decimal, ParameterDirection.Input);
             parameter.Add("@CoversWinPercentage", team.CoversWinPercentage, DbType.Int32, ParameterDirection.Input);
 
             var id = connection.ExecuteScalar("spInsertTeam",
